dotaguides/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.3
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.2.0
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
